7mo avance II

+++ app/src/main/java/com/example/pruebatalana/ui/view/HomeActivity.kt
@@ -3,6 +3,7 @@ package com.example.pruebatalana.ui.view
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.pruebatalana.PruebaTalanaApp.Companion.prefs
@@ -29,8 +30,17 @@ class HomeActivity : AppCompatActivity() {
         val userName = prefs.getUser()
         homeBinding.tvUser.text="Bienvenido $userName "
         homeBinding.btnCerrarSesion.setOnClickListener {
-            prefs.clear()
-            onBackPressed()
+            val builder = AlertDialog.Builder(this)
+            builder.setTitle("¿Seguro que desea cerrar sesión?")
+                .setMessage("Si continua, deberá ingresar nuevamente con sus credenciales")
+                .setPositiveButton("Si"){ _, it->
+                    prefs.clear()
+                    onBackPressed()
+                }
+                .setNegativeButton("No",null )
+                .create()
+                .show()
+
         }
 
     }
